Post-morten BSM: Reminders from BSM mostly regarding messaging stuff. (#182)
* some clean ups

* contiguous secuences

* add UpdateHoldingValue msg

* remove categories

* minor rename encodePacked() to toBytes()

* rename TriggerUpdateHolding to TriggerUpdateHoldingAmount

* Use plain types

* timestamp as u64

* address to bytes32

* move BSM test file to integration

* Remove deployer access

* minor simplification

* timestamps in MessageDispatcher

* remove increase/decrease holdings

* fix stack too deep issue

* add tests for JournalEntry

* tests: improve D18 & MultiShareClass coverage post BSM (#188)

* chore: add missing D18 coverage

* docs: update com

* tests: add multishareclass cov

* remove old snap file

* add a comment

* move JournalEntryLib to libraries

---------

+++ test/vaults/integration/SyncDeposit.t.sol
@@ -14,7 +14,7 @@ import {IGateway} from "src/common/interfaces/IGateway.sol";
 import {PoolId} from "src/common/types/PoolId.sol";
 import {ShareClassId} from "src/common/types/ShareClassId.sol";
 import {AssetId} from "src/common/types/AssetId.sol";
-import {JournalEntry} from "src/common/types/JournalEntry.sol";
+import {JournalEntry} from "src/common/libraries/JournalEntryLib.sol";
 
 import {IBalanceSheetManager} from "src/vaults/interfaces/IBalanceSheetManager.sol";
 import {SyncDepositVault} from "src/vaults/SyncDepositVault.sol";
@@ -108,7 +108,7 @@ contract SyncDepositTest is BaseTest {
             syncRequests.escrow(),
             depositAssetAmount,
             pricePerUnit,
-            timestamp,
+            uint64(timestamp),
             journalEntries,
             journalEntries
         );
@@ -117,7 +117,7 @@ contract SyncDepositTest is BaseTest {
         emit IGateway.SendMessage(bytes(""));
         vm.expectEmit();
         emit IBalanceSheetManager.UpdateValue(
-            poolId, scId, vault.asset(), vaultDetails.tokenId, pricePerUnit, timestamp
+            poolId, scId, vault.asset(), vaultDetails.tokenId, pricePerUnit, uint64(timestamp)
         );
     }
 }
